Import WireFormat qualified
Enough messing about worrying about name collisions. By using a
separate module, Receiver, to be the [only] place where both CoreTypes
and WireFormat meet, we can stop worrying about name collisions
elsewhere.

In this source file, switch to importing WireFormat qualified as well.
It worked with only one of them imported qualified, but for all that
its more cumbersome it is explicitly very clear which is being used on
each side of the conversion functions.