Preliminary infrastructure to update contents list
Begin groundwork to update contents list. Have appendVaultSource
return a Set of the sources we didn't already know about, and then
after returning the ack via zeromq, send off the write to append to
the contents object.

We now have as a given that a DataBurst represents points from a
single origin. So drop the code which grouped Points by origin

Change traverseWithKey (which rebuilds map) to a fold. Thanks again to
@christian-marie for his help on getting the types right.