Fix ingestd appearing to hang on zmq exceptions
The zeromq-haskell library's `Zero.async` has a `finally` call in it,
catching all exceptions and decrementing the reference count held by
`runZMQ`. The problem with this is that it meant that exceptions were
not propegating out of that thread and thereby not taking out all the
linked ones so the process was hung, effectively.

Fix by switching to using `Async.waitEitherCancel`