Refactor Address to be a newtype wrapper
There are far too many Word64 floating around, and getting the
argument order wrong is starting to be a problem. A newtype wrapper is
the obvious answer, with the usual baggage.

In this case, moved the bit fiddling operations to Vaultaire.CoreTypes
in an attempt to do it in one place rather than several.